require and rename
